+++ CHANGES.md
@@ -8,7 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ### Fixed
   - Setting KillApps in walkthru mode now shows a prompt for each line expecting input.
-  - Now correctly differentiates `false` from `nil` values when updating a titles changelog.
+  - Now correctly differentiates `false` from `nil` values when updating a title's changelog.
  
 ## \[1.0.1] - 2025-10-02